Reconstruct animation and face behaviour from the engine where it had been inferred
Animation layer (AnimationScheduler, CozmoAnimations, AnimationLibrary):
- Head and lift keyframes stream as animHeadAngle/animLiftHeight (0x93/0x94) with the keyframe's
  variability applied at stream time via RandIntInRange (HeadAngleKeyFrame::GetStreamMessage 0x004F8C08,
  LiftHeightKeyFrame 0x004F8F80), not as SetHeadAngle/SetLiftHeight motor commands.
- Audio alternatives are chosen by cumulative probability from a single draw, 1/n when the
  probability list is absent or mismatched (RobotAudioKeyFrame::GetAudioRefIndex 0x004F9AEC,
  SetMembersFromFlatBuf 0x004F9E54). The first-producible-alternative fallback is labelled local.
- AnimationLibrary.Choose records the engine's weighted selection, cooldown and UseHeadAngle gate
  (AnimationGroup::GetAnimationName 0x0058A970); enforcement stays deferred.

Procedural face (ProceduralFace, Expressions, ProceduralFaceRenderer, IdleBehavior):
- Interpolate blends eye and face angles as directions and clips parameters with the engine's table
  (ProceduralFace::Interpolate 0x00584290, Clip 0x005847A8, table 0x00C5A97C); face scales floor at 0.
- The resting face is the shipped anim_neutral_eyes_01 pose that AnimationStreamer (0x00579F78)
  installs from AnimationTrigger::NeutralFace, not a symmetric canvas-centre pose. Renderer.Neutral()
  becomes Nominal() to mark it as a measurement pose only.
- Blinks follow the seven-frame scale table at 0x00C5AAD8 (GetNextBlinkFrame 0x00585F18), 331 ms,
  multiplied onto the base pose as a layer (Combine 0x005846A8).
- Darts follow GenerateEyeShift 0x0058D100 / LookAt 0x00584158: whole-face move of x,y in ±6, EyeScaleY
  only, convergence when looking down; the min/max dart scale clamp is removed because LookAt never
  reads it. Dart hold-and-return lifecycle remains local policy and is labelled unresolved.

Tests: SourceFidelityTests added; AnimationGapTests, IdleFaceTests, DeviceTests and recorder fixtures
updated for the new sink signature and behaviour. 440 tests pass.

+++ cozmo-stack/tests/Cozmo.Protocol.Tests/HardeningTests.cs
@@ -34,8 +34,8 @@ public class HardeningTests
         private readonly object _gate = new();
         public void Face(FaceBitmap b) { }
         public void Audio(byte[]? f) { }
-        public void Head(float r, uint d) { }
-        public void Lift(float h, uint d) { }
+        public void Head(sbyte r, uint d) { }
+        public void Lift(byte h, uint d) { }
         public void AnimationStarted(byte tag) { }
         public void AnimationEnded() { }
         public void Body(BodyKeyframe k) { }
